% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrigeMultiplosCedears.R
\name{corrigeMultiplosCedears}
\alias{corrigeMultiplosCedears}
\title{corrigeMultiplosCedears}
\usage{
corrigeMultiplosCedears(df)
}
\arguments{
\item{df}{el dataframe que devuelve getPPIPriceHistoryMultiple3}
}
\value{
Devuelve el mismo df pero con el campo price de los cedears ajustados.
}
\description{
Función que corrige los múltiplos de los cedears hacia atrás
No tiene todos los CEDEARs. A medida que salen nuevos hay que ir
incluyéndolos
Hay que pasarle el df que vuelve de PPI
Lo devuelve con los precios (columna "price") ajustados
}
