% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPPICCL.R
\name{getPPICCL}
\alias{getPPICCL}
\title{getPPICCL}
\usage{
getPPICCL(from = Sys.Date(), to = Sys.Date(), via = c("AAPL", "AAPLD"))
}
\arguments{
\item{from}{Fecha desde. Default hoy}

\item{to}{Fecha hasta. Default hoy}

\item{via}{Nombre CEDEAR a usar. Default AAPL, AAPLD. Primero en pesos.}
}
\value{
Un tibble con el ccl
}
\description{
Devuelve el CCL a través de CEDEARs en PPI
}
