% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsPPI.R
\name{getPPICurrentRofex}
\alias{getPPICurrentRofex}
\title{getPPICurrentRofex}
\usage{
getPPICurrentRofex(db = "", spt = NULL)
}
\arguments{
\item{db}{Base donde va a buscar los feriados. Default es test.}

\item{spt}{Spot para usar. Overrides el de MAE. Si no se pone nada y MAE no lo tiene devuelve error.}
}
\value{
una lista con un df con los futuros y un gráfico con las implícitas
}
\description{
Trae la curva de Rofex desde BYMA (via PPI) y calcula implícitas
Toma el spot de la web de MAE. Si se le informa uno en spt, toma ese.
}
