% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsPPI.R
\name{getPPIDLR}
\alias{getPPIDLR}
\title{Devuelve el histórico del precio del dolar MEP y CCL y el gráfico.}
\usage{
getPPIDLR(from = "2014-05-27", to = Sys.Date(), settlement = "t+0")
}
\arguments{
\item{from}{Fecha desde la que buscar. Default es 2014-05-27}

\item{to}{Fecha hasta la que buscar. Default es Sys.Date()}

\item{settlement}{Tipo de liquidación. Default es "t+0"}
}
\value{
Un tibble con los valores de dólar
}
\description{
Devuelve el histórico del precio del dolar MEP y CCL y el gráfico. Lo calcula con GD30 y AL30 desde septiembre 2020. Si se pide fecha previa, lo construye con AY24. La fecha más atrás que posee es 2014-05-27
}
\details{
getPPIDLR
Trae valores de dólar de PPI
}
