\name{getPPILogin}
\alias{getPPILogin}
\title{Login de conexión con la API de PPI. (DEPRECADA)}
\usage{
getPPILogin()
}
\description{
Obtiene el token para utilizar con la API de PPI.
Está deprecada. Fue reemplazada por getPPILogin2

Devuelve una lista con dos elementos.

1.- El token (bearer)
2.- El refresh token
}
\examples{
getPPILogin()
}

