% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsPPI.R
\name{getPPIPrices}
\alias{getPPIPrices}
\title{getPPIPrices}
\usage{
getPPIPrices(token, ticker, type, from, to, settlement = "INMEDIATA", ...)
}
\arguments{
\item{token}{Token de PPI}

\item{ticker}{Tickers a buscar. Acepta vector.}

\item{type}{Tipo de activo. Siempre usar único tipo por vez. Acepta vector.}

\item{from}{Fecha desde la que buscar}

\item{to}{Fecha hasta la que buscar}

\item{settlement}{Tipo de liquidación. Default es "INMEDIATA"}

\item{...}{Parámetros para dbExecuteQuery}
}
\value{
Lista con dos elementos: result y fail. Result es un tibble con los precios encontrados y fail los tickers que no se encontraron.
}
\description{
Trae los precios de un ticker en un rango de fechas, y si PPI no lo tiene
lo busca en la base de datos local.
Esta es por ahora la más completa al 7-sep-2024 ya que es la primera que incluye
poder buscar en base de datos cuando PPI no tiene los precios.
}
